refactor!: base_images_dockerfile -> sbom_utils
